+++ src/imio/migrator/migrator.py
@@ -154,3 +154,10 @@ class Migrator:
             # the profile is well installed
             if profile not in omit and self.ps.getLastVersionForProfile(profile) != 'unknown':
                 self.upgradeProfile(profile)
+
+    def runProfileSteps(self, product, steps=[], profile='default'):
+        """ Run given steps of a product profile (default is 'default' profile) """
+        for step_id in steps:
+            logger.info("Running profile step '%s:%s' => %s" % (product, profile, step_id))
+            self.portal.portal_setup.runImportStepFromProfile('profile-%s:%s' % (product, profile), step_id)
+